refactor: split static property codegen
